Added split generation functions in a separate file
